Resizing now changes draggable constraints.
The containment of the draggable part of things is now tied to
centerMap, meaning it'll always readjust itself if the window gets
resized for any reason.

Note that this still doesn't account for displays larger than 3200x1600,
but since we don't have any like that here at Lexmark, I think that's
acceptable for the time being.