Added opacifying to the forecast box.
Well, the forecast box already had a slightly-transparent background,
but now it'll turn the entire thing, text and all, transparent while
dragging the map around.  What's more, it'll do so with a quick CSS3
transision!

Not that it'll do us much good on the computer on which Lexstar runs,
but it looks nifty in other uses.