Changed all em measurements to px.
Using em measurements may be nice and all, but when you have to
integrate them with images that are most definitely not in em
measurements, things start to get hairy, especially when the definition
of "em" might change between browsers or configurations.  So, all em
stuff is now px, at a rate of 1em = 16px.