Added icon picking.
I'm just recycling the same icons I have now, rather than attempt to
replicate all of the NOAA's forecast icons, mostly because those
include X% chance-of-whatever stuff, which I'm not in the mood to sort
out right now.  Instead, it'll just base it off of the condition that
got prioritized earlier.