Made the forecast data work.
There it goes.  The forecast data now shows up properly when the button
gets clicked.  This is some of the sloppiest code I've written, and it
gets grossly inconsistent with the old code, but it works.

So I guess I should fix all the code next, whenever I get to it.